ERA5Reanalysis

Documentation for ERA5Reanalysis.

ERA5Reanalysis.ERA5Dataset

ERA5Reanalysis.ERA5Variable

ERA5Reanalysis.cdskey

ERA5Reanalysis.isPressureVariable

ERA5Reanalysis.isSingle

ERA5Reanalysis.listPressureVariables

ERA5Reanalysis.listSingles

ERA5Reanalysis.parserequest

ERA5Reanalysis.retrieve

ERA5Reanalysis.ERA5Dataset — Type

ERA5Dataset

Abstract supertype for ERA5 datasets.

ERA5Reanalysis.ERA5Variable — Type

" ERA5Dataset

Abstract supertype for ERA5 datasets.

ERA5Reanalysis.cdskey — Method

cdskey() -> Dict{Any,Any}

Retrieves the CDS API ckeys from the ~/.cdsapirc file in the home directory

ERA5Reanalysis.isPressureVariable — Method

isPressureVariable(
    varID :: AbstractString;
    throw :: Bool = true
) -> tf :: Bool

Extracts information of the Pressure-Level Variable with the ID varID. If no Pressure-Level Variable with this ID exists, an error is thrown.

Arguments

RegID : The keyword ID that will be used to identify the Pressure-Level Variable. If the ID is not valid (i.e. not being used), then an error will be thrown.

throw : If true, then throws an error if RegID is not a valid Pressure-Level Variable identifier instead of returning the Boolean tf

Returns

tf : True / False

ERA5Reanalysis.isSingle — Method

isSingle( varID :: AbstractString; throw :: Bool = true ) -> tf :: Bool

Extracts information of the Single-Level Variable with the ID varID. If no Single-Level Variable with this ID exists, an error is thrown.

Arguments

RegID : The keyword ID that will be used to identify the Single-Level Variable. If the ID is not valid (i.e. not being used), then an error will be thrown.

throw : If true, then throws an error if RegID is not a valid Single-Level Variable identifier instead of returning the Boolean tf

Returns

tf : True / False

ERA5Reanalysis.listPressureVariables — Method

listPressureVariables()

List all Pressure-Level Variables and the files the data are stored in.

Output

varlist : List of all the Pressure-Level Variable IDs

fidlist : List of the files that the Pressure-Level Variable information is stored in

ERA5Reanalysis.listSingles — Method

listSingles()

List all Single-Level Variables and the files the data are stored in.

Output

varlist : List of all the Single-Level Variable IDs

fidlist : List of the files that the Single-Level Variable information is stored in

ERA5Reanalysis.parserequest — Method

parserequest(
    ckeys :: AbstractDict,
    resp  :: AbstractDict,
    api   :: AbstractString
) -> Dict{Any,Any}

Get info on HTTP request, and parse the information and update the dictionary

Arguments

ckeys : Dictionary that contains the CDSAPI information held in ~/.cdsapirc

resp : Dictionary that contains the HTTP response

api : String that contains the API used for Climate Data Store authentication

ERA5Reanalysis.retrieve — Function

retrieve(
    dset  :: AbstractString,
    dkeys :: AbstractDict,
    fnc   :: AbstractString,
    ckeys :: AbstractDict = cdskey()
)

Retrieves dsets from the Climate Data Store, with options specified in a Julia Dictionary and saves it into a specified file. Arguments: * dset::AbstractString : string specifies the name of the dset within the Climate Data Store that the retrieve function is attempting to retrieve data from * dkeys::AbstractDict : dictionary that contains the keywords that specify the properties (e.g. date, resolution, grid) of the data being retrieved * fnc::AbstractString : string that contains the path and name of the file that the data is to be saved into * ckeys::AbstractDict : dictionary that contains API Key information read from the .cdsapirc file in the home directory (optional)