The Basic Components of ERA5Reanalysis.jl

There are three essential components to specifying an ERA5 reanalysis dataset:

The ERA5 module (hourly, monthly, month-by-hour, etc.), held in an ERA5Module supertype

The ERA5 variable (single-level, or pressure-level), held in the ERA5Variable supertype

The ERA5 region to be downloaded/analyzed, held in an ERA5Region supertype, built over a GeoRegion

The ERA5Region superType

Finally, in order to specify an area for ERA5Reanalysis.jl to conduct downloads, or analyse data over, we need to specify an ERA5Region, which is built on top of the GeoRegion Type in GeoRegions.jl.

ERA5Reanalysis.ERA5Region — Type

ERA5Region

Structure that imports relevant GeoRegion properties used in the handling of the gridded ERA5 datasets.

All ERA5Region Types contain the following fields:

geo : The GeoRegion containing the geographical information

geoID : The ID used to specify the GeoRegion

gres : The resolution of the gridded data to be downloaded/analysed

fstr : String, for specification of folder and file name

isglb : A Bool, true if spans the globe, false if no