The ERA5Dataset superType

ERA5 reanalysis data is stored on the Climate Data Store in several different categories, so different people with different needs may access different data types depending on their research. In ERA5Reanalysis, we defined these datasets as ERA5Dataset Types.

When defining an ERA5Dataset container, we also indicate the start and end dates of the dataset that we want to work on.

ERA5Reanalysis.ERA5Dataset — Type

ERA5Dataset

Abstract supertype for ERA5 reanalysis datasets, with the following subtypes:

ERA5Hourly  <: ERA5Dataset
ERA5Daily   <: ERA5Dataset
ERA5Monthly <: ERA5Dataset

All ERA5Dataset Types contain the following fields:

ID : The module ID, that also acts as a prefix to filenames

name : The full name of the module

start : The date for which downloads/analysis begins

stop : The date for which downloads/analysis finishes

path : The specified directory in which to save the data

emask : The specified directory in which to save the LandSea dataset

sldoi : Single-Level DOI (N/A for ERA5Daily)

pldoi : Pressure-Level DOI (N/A for ERA5Daily)

ptype : Product type (N/A for ERA5Daily), set to reanalysis

The ERA5Monthly Type will also contain the following fields:

hours : specifies the hour(s) of day for which monthly data is downloaded

Available in the Climate Data Store (CDS)

The following datasets are currently available by default in the Climate Data Store:

Hourly reanalysis data (ERA5Hourly)

Monthly reanalysis data, (ERA5Monthly)

Note

The Climate Data Store also has an option to download ERA5 Monthly data by hour-of-day. This is subsumed into the ERA5Monthly dataset and can be selected by specifying hours = true.

ERA5Reanalysis.ERA5Hourly — Type

ERA5Hourly <: ERA5Dataset

Specifies that the dataset to be analyzed contains hourly data. All fields are the same as that specified in the ERA5Dataset docstring.

ERA5Reanalysis.ERA5Hourly — Method

ERA5Hourly(;
    start :: TimeType,
    stop  :: TimeType,
    path  :: AbstractString = homedir(),
) -> ERA5Hourly <: ERA5Dataset

A function that creates an ERA5Hourly module. All possible hours are downloaded, and data is saved month-by-month.

Keyword Arguments

path : The specified directory in which to save the data

start : The date for which downloads/analysis begins, automatically rounded to the nearest month

stop : The date for which downloads/analysis finishes, automatically rounded to the nearest month

ERA5Reanalysis.ERA5Monthly — Type

ERA5Monthly <: ERA5Dataset

Specifies that the dataset to be analyzed contains monthly-mean data. All fields are the same as that specified in the ERA5Dataset docstring.

ERA5Reanalysis.ERA5Monthly — Method

ERA5Monthly(;
    start :: TimeType,
    stop  :: TimeType,
    path  :: AbstractString = homedir(),
    hours :: Bool = false,
) -> ERA5Monthly <: ERA5Dataset or ERA5MonthlyHour <: ERA5Dataset

A function that creates an ERA5Monthly or ERA5MonthlyHour module depending on the input arguments of hours. Data is saved year-by-year.

Keyword Arguments

path : The specified directory in which to save the data

start : The date for which downloads/analysis begins, automatically rounded to the nearest year

stop : The date for which downloads/analysis finishes, automatically rounded to the nearest year

hours : If false, download monthly-averaged data. If true, download monthly-averaged data for each hour

Custom ERA5Reanalysis.jl Dataset, not available in CDS

We also have created a custom ERA5Daily dataset, which is a daily-average of the hourly data from ERA5Hourly data.

ERA5Reanalysis.ERA5Daily — Type

ERA5Daily <: ERA5Dataset

Specifies that the dataset to be analyzed contains hourly data. All fields are the same as that specified in the ERA5Dataset docstring.

Warning

Because the Climate Data Store does not provide daily data by default, the download() functionality cannot be used for this ERA5Dataset type.

There are other potential modules that could be incorporated into ERA5Reanalysis.jl, such as

ERA5-Land data

Ensemble model averages, individual members, and standard deviations