ERA5 Datasets available in CDS

The ERA5CDStore Type of the parent type of all ERA5 Datasets that are naturally available from the Climate Data Store and can thus be downloaded directly using the CDSAPI.

Note

The Climate Data Store also has an option to download ERA5 Monthly data by hour-of-day. This is subsumed into the ERA5Monthly dataset and can be selected by specifying hours = true.

ERA5Reanalysis.ERA5Hourly — Type

ERA5Hourly <: ERA5CDStore

Specifies that the dataset to be analyzed contains hourly data. All fields are the same as that specified in the ERA5Dataset docstring.

ERA5Reanalysis.ERA5Monthly — Type

ERA5Monthly <: ERA5CDStore

Specifies that the dataset to be analyzed contains monthly-mean data. The ERA5Monthly Type will also contain the following fields:

hours : specifies the hour(s) of day for which monthly data is downloaded

There are other potential modules that could be incorporated into ERA5Reanalysis.jl, such as

ERA5-Land data

Ensemble model averages, individual members, and standard deviations

They have not been added yet into ERA5Reanalysis.jl. If you are potentially interested in me adding these datasets to ERA5Reanalysis.jl, please submit a pull request.

API

ERA5Reanalysis.ERA5Hourly — Method

ERA5Hourly(;
    start :: TimeType,
    stop  :: TimeType,
    path  :: AbstractString = homedir(),
) -> ERA5Hourly <: ERA5Dataset

A function that creates an ERA5Hourly module. All possible hours are downloaded, and data is saved month-by-month.

Keyword Arguments

path : The specified directory in which to save the data

start : The date for which downloads/analysis begins, automatically rounded to the nearest month

stop : The date for which downloads/analysis finishes, automatically rounded to the nearest month

ERA5Reanalysis.ERA5Monthly — Method

ERA5Monthly(;
    start :: TimeType,
    stop  :: TimeType,
    path  :: AbstractString = homedir(),
    hours :: Bool = false,
) -> ERA5Monthly <: ERA5Dataset or ERA5MonthlyHour <: ERA5Dataset

A function that creates an ERA5Monthly or ERA5MonthlyHour module depending on the input arguments of hours. Data is saved year-by-year.

Keyword Arguments

path : The specified directory in which to save the data

start : The date for which downloads/analysis begins, automatically rounded to the nearest year

stop : The date for which downloads/analysis finishes, automatically rounded to the nearest year

hours : If false, download monthly-averaged data. If true, download monthly-averaged data for each hour