Custom ERA5 Datasets defined in ERA5Reanalysis.jl

We also have created a custom ERA5Daily dataset, which is a daily-average of the hourly data from ERA5Hourly data.

ERA5Reanalysis.ERA5Daily — Type

ERA5Daily <: ERA5Custom

Specifies that the dataset to be analyzed contains hourly data. All fields are the same as that specified in the ERA5Dataset docstring. However, the fields sldoi, pldoi and ptype are all set to N/A because ERA5Daily is not available in, nor downloadable from, the Climate Data Store.

ERA5Reanalysis.ERA5Daily — Method

ERA5Daily(;
    start :: TimeType,
    stop  :: TimeType,
    path  :: AbstractString = homedir(),
) -> ERA5Daily <: ERA5Dataset

A function that creates an ERA5Hourly module. All possible hours are downloaded, and data is saved month-by-month.

Keyword Arguments

path : The specified directory in which to save the data

start : The date for which downloads/analysis begins, automatically rounded to the nearest month

stop : The date for which downloads/analysis finishes, automatically rounded to the nearest month

There are other potential modules that could be incorporated into ERA5Reanalysis.jl, such as

ERA5-Land data

Ensemble model averages, individual members, and standard deviations

API