Dummy ERA5 Datasets

Dummy datasets are meant to specify the paths to the ERA5 data root directory, and LandSea folder, without needing inputs for date.

ERA5Reanalysis.ERA5Dummy — Type

ERA5Dummy <: ERA5Dataset

The ERA5Dummy dataset contains only information on the data and mask paths.

julia> using ERA5Reanalysis
julia> e5ds = ERA5Dummy(path=homedir())[ Info: 2025-11-29T00:33:14.282 - ERA5Reanalysis.jl - Setting up data structure for the ERA5 Dummy Dataset
The ERA5Monthly Module has the following properties:
    Mask Directory (emask) : /home/runner/emask
julia> isdir(e5ds.emask)true

API

ERA5Reanalysis.ERA5Dummy — Method

ERA5Dummy(;
    path  :: AbstractString = homedir(),
) -> ERA5Dummy <: ERA5Dataset

A function that creates a dummy ERA5Dataset that contains only information on the path of the ERA5 LandSea mask

Keyword Arguments

path : The specified directory in which to save the data