Creating Custom ERA5Variables

While the list of available ERA5Variables in the Climate Data Store is very extensive, they are not exhaustive. Furthermore, due to the very large number of SingleLevel variables, we only predefine a small subset of the variables available on CDS in ERA5Reanalysis.jl. There is therefore a need to define a list of custom ERA5Variables that will be otherwise calculated (although this has to be done separately outside of ERA5Reanalysis.jl), and then can thus be saved and manipulated and analyzed in the same way that the default ERA5Variables in ERA5Reanalysis.jl can be. In this page, we break this down into the SingleLevel and PressureLevel components

Custom SingleLevel Variables

We allow the user the option to define if the custom SingleLevel variable they are defining exists on the Climate Data Store, and store it as a SingleVariable type. If a variable is not available on the Climate Data Store, they should be defined as a SingleCustom type.

Defining a new SingleVariable or SingleCustom

To define both SingleVariable and SingleCustom variables, we use the funciton SingleVariable(). To create a SingleCustom variable, the keyword argument inCDS must be set to false.

ERA5Reanalysis.SingleVariable — Type

SingleVariable(
    ST = String;
    ID :: AbstractString,
    long :: AbstractString = "",
    name :: AbstractString,
    units :: AbstractString,
    inCDS :: Bool = true
) -> evar :: SingleLevel

Create a custom Single-Level variable that is not in the default list exported by ERA5Reanalysis.jl. These variables are either available in the CDS store (whereby they can be both downloaded analyzed), or not (in which case means that they were separately calculated from other variables and analyzed).

Keyword Arguments

ID : variable ID (in string format) used in the NetCDF file

long : long-name for variable (used in specifying variable for CDS downloads)

name : user-defined variable name

units : user-defined units of the variable

inCDS : Boolean that indicates if this variable is available on the CDS store. True if available.

Removing Custom SingleLevel Variables

To remove a SingleLevel variable, we can use the rmSingle() function:

ERA5Reanalysis.rmSingle — Method

rmSingle( ID :: AbstractString ) -> nothing

Remove the Single-Level Variable with the ID ID from the lists.

Arguments

RegID : The keyword ID that will be used to identify the Single-Level Variable that is to be removed

An Example!

In this example, we add two SingleLevel variables to the list:

hvc, for High Vegetation Cover, which is a downloadable variable in the CDS

eke, for Eddy Kinetic Energy, which is not downloadable and must be calculated using other variables

The Eddy Kinetic Energy is calculated by the following:

\[E = \frac{1}{2g} \int_0^{p_s} \overline{u'^2} + \overline{v'^2} \>dp\]

And it is often used as a measure of storm-track intensity.

julia> using ERA5Reanalysis
julia> SingleVariable(
           varID = "cvh",
           units = "0-1",
           vname = "High Vegetation Cover",
           lname = "high_vegetation_cover",
       )ERROR: UndefKeywordError: keyword argument `ID` not assigned
julia> SingleVariable(
           varID = "eke",
           units = "J m**-2",
           vname = "Eddy Kinetic Energy",
           lname = "eddy_kinetic_energy",
           inCDS = false
       )ERROR: UndefKeywordError: keyword argument `ID` not assigned
julia> tableSingles() ---------------- ------ -------------------------------- ----------- ---------------------------------------
  Variable Type     ID    Name                               Units     ERA5 Long-Name
 ---------------- ------ -------------------------------- ----------- ---------------------------------------
  SingleVariable   u10    Zonal Wind at 10m Height          m s**-1    10m_u_component_of_wind
  SingleVariable   v10    Meridional Wind at 10m Height     m s**-1    10m_v_component_of_wind
  SingleVariable   u100   Zonal Wind at 100m Height         m s**-1    100m_u_component_of_wind
  SingleVariable   v100   Meridional Wind at 100m Height    m s**-1    100m_v_component_of_wind
  SingleVariable   t2m    2m Temperature                       K       2m_temperature
  SingleVariable   d2m    2m Dewpoint Temperature              K       2m_dewpoint_temperature
  SingleVariable   sst    Sea Surface Temperature              K       sea_surface_temperature
  SingleVariable   skt    Skin Temperature                     K       skin_temperature
  SingleVariable    sp    Surface Pressure                    hPa      surface_pressure
  SingleVariable   msl    Sea Level Pressure                  hPa      mean_sea_level_pressure
  SingleVariable    tp    Total Precipitation                  m       total_precipitation
  SingleVariable    cp    Convective Precipitation             m       convective_precipitation
  SingleVariable   lsp    Large-Scale Precipitation            m       large_scale_precipitation
  SingleVariable   tcw    Total Column Water               kg m**-2    total_column_water
  SingleVariable   tcwv   Total Column Water Vapour        kg m**-2    total_column_water_vapour
  SingleVariable   tcc    Total Cloud Cover                   0-1      total_cloud_cover
  SingleVariable   hcc    High Cloud Cover                    0-1      high_cloud_cover
  SingleVariable   mcc    Medium Cloud Cover                  0-1      medium_cloud_cover
  SingleVariable   lcc    Low Cloud Cover                     0-1      low_cloud_cover
  SingleVariable   cape   CAPE                             J kg**-1    convective_available_potential_energy
  SingleVariable   cin    CIN                              J kg**-1    convective_inhibition
  SingleVariable   tsr    Top Net Solar Radiation           J m**-2    top_net_solar_radiation
  SingleVariable   ttr    Top Net Thermal Radiation         J m**-2    top_net_thermal_radiation
  SingleVariable   ssr    Surface Net Solar Radiation        J m**-2   surface_net_solar_radiation
  SingleVariable   str    Surface Net Thermal Radiation     J m**-2    surface_net_thermal_radiation
  SingleVariable   sshf   Surface Sensible Heat Flux        J m**-2    surface_sensible_heat_flux
  SingleVariable   slhf   Surface Latent Heat Flux          J m**-2    surface_latent_heat_flux
 ---------------- ------ -------------------------------- ----------- ---------------------------------------
julia> rmSingle.(["eke","hvc"])┌ Warning: 2023-06-03T04:20:25.416 - ERA5Reanalysis.jl - "eke" is not a valid SingleVariable identifier, use the function SingleVariable() or SingleCustom() to add this ERA5Variable to the list.
└ @ ERA5Reanalysis ~/work/ERA5Reanalysis.jl/ERA5Reanalysis.jl/src/variable/single.jl:198
┌ Warning: 2023-06-03T04:20:25.416 - ERA5Reanalysis.jl - No Single-Level variable defined by "eke" exists, please make sure you specified the correct variable ID
└ @ ERA5Reanalysis ~/work/ERA5Reanalysis.jl/ERA5Reanalysis.jl/src/variable/single.jl:228
┌ Warning: 2023-06-03T04:20:25.417 - ERA5Reanalysis.jl - "hvc" is not a valid SingleVariable identifier, use the function SingleVariable() or SingleCustom() to add this ERA5Variable to the list.
└ @ ERA5Reanalysis ~/work/ERA5Reanalysis.jl/ERA5Reanalysis.jl/src/variable/single.jl:198
┌ Warning: 2023-06-03T04:20:25.417 - ERA5Reanalysis.jl - No Single-Level variable defined by "hvc" exists, please make sure you specified the correct variable ID
└ @ ERA5Reanalysis ~/work/ERA5Reanalysis.jl/ERA5Reanalysis.jl/src/variable/single.jl:228
2-element Vector{Nothing}:
 nothing
 nothing
julia> isSingle("eke",throw=false) # don't throw error, just show warning┌ Warning: 2023-06-03T04:20:25.905 - ERA5Reanalysis.jl - "eke" is not a valid SingleVariable identifier, use the function SingleVariable() or SingleCustom() to add this ERA5Variable to the list.
└ @ ERA5Reanalysis ~/work/ERA5Reanalysis.jl/ERA5Reanalysis.jl/src/variable/single.jl:198
false

Custom PressureLevel variables

In ERA5Reanalysis.jl, we have predefined all the pressure-level variables that are available directly from the CDS. As such, all custom PressureLevel variables are defined under the PressureCustom type.

Defining Custom PressureLevel Variables

PressureCustom variables are defined using the PressureVariable() function, as follows:

ERA5Reanalysis.PressureVariable — Type

PressureVariable(
    ST = String;
    ID :: AbstractString,
    long :: AbstractString = "",
    name :: AbstractString,
    units :: AbstractString,
    hPa   :: Int = 0,
    throw :: Bool = true
) -> evar :: PressureCustom

Create a custom Pressure-Level variable that is not in the default list exported by ERA5Reanalysis.jl. These variables are not available in the CDS store, and so they must be separately calculated from other variables and analyzed.

Keyword Arguments

ID : variable ID (in string format) used in the NetCDF file

long : long-name for variable (used in specifying variable for CDS downloads)

name : user-defined variable name

units : user-defined units of the variable

hPa : Pressure level specified in hPa. Default is 0, which indicates all levels.

throw : if hPa level does not exist and throw is true, throw error, otherwise find nearest pressure level

Removing Custom PressureLevel Variables

To remove a PressureCustom variable, we can use the rmPressure() function:

ERA5Reanalysis.rmPressure — Method

rmPressure( ID :: AbstractString ) -> nothing

Remove the Pressure-Level Variable with the ID ID from the lists.

Arguments

RegID : The keyword ID that will be used to identify the Pressure-Level Variable that is to be removed

An Example!

Let us define the custom variable "vt" for Virtual Temperature, which is a proxy to buoyancy in the atmosphere. Virtual Temperature is calculated from both the water vapour mixing ratio and atmospheric temperature and is given by the formula:

\[T_v = T(1+0.61q)\]

Virtual Temperature has the units "K", and for this example let us define the pressure level height we are interested in to be at 1000 hPa. The resultant PressureCustom variable is given by:

julia> using ERA5Reanalysis
julia> PressureVariable(
           varID = "vt",
           units = "K",
           hPa   = 1000,
           lname = "virtual_temperature",
           vname = "Virtual Temperature"
       )ERROR: UndefKeywordError: keyword argument `ID` not assigned
julia> tablePressures() ------------------ ------ ------------------------------------- ---------------------- -------------------------------------
   Variable Type      ID    Name                                         Units           ERA5 Long-Name
 ------------------ ------ ------------------------------------- ---------------------- -------------------------------------
  PressureVariable    u     Zonal Wind                                  m s**-1          u_component_of_wind
  PressureVariable    v     Meridional Wind                             m s**-1          v_component_of_wind
  PressureVariable    w     Vertical Wind                               Pa s**-1         vertical_velocity
  PressureVariable    t     Temperature                                    K             temperature
  PressureVariable    z     Pressure Level Geopotential                m**2 s**-2        geopotential
  PressureVariable    d     Divergence                                   s**-1           divergence
  PressureVariable    vo    Vorticity (Relative)                         s**-1           vorticity
  PressureVariable    pv    Potential Vorticity                   K m**-2 kg**-1 s**-1   potential_vorticity
  PressureVariable    r     Relative Humidity                              %             relative_humidity
  PressureVariable    q     Specific Humidity                          kg kg**-1         specific_humidity
  PressureVariable    cc    Cloud Cover Fraction                           %             fraction_of_cloud_cover
  PressureVariable   ciwc   Specific Cloud Ice Water Content           kg kg**-1         specific_cloud_ice_water_content
  PressureVariable   clwc   Specific Cloud Liquid Water Content        kg kg**-1         specific_cloud_liquid_water_content
  PressureVariable   crwc   Specific Rain Water Content                kg kg**-1         specific_rain_water_content
  PressureVariable   cswc   Specific Snow Water Contnet                kg kg**-1         specific_snow_water_content
  PressureVariable    o3    Ozone Mass Mixing Ration                   kg kg**-1         ozone_mass_mixing_ratio
 ------------------ ------ ------------------------------------- ---------------------- -------------------------------------
julia> rmPressure("vt")┌ Warning: 2023-06-03T04:20:26.570 - ERA5Reanalysis.jl - "vt" is not a valid PressureVariable identifier, use the function PressureCustom() to add this ERA5Variable to the list.
└ @ ERA5Reanalysis ~/work/ERA5Reanalysis.jl/ERA5Reanalysis.jl/src/variable/pressure.jl:233
┌ Warning: 2023-06-03T04:20:26.571 - ERA5Reanalysis.jl - No Pressure-Level variable defined by "vt" exists, please make sure you specified the correct variable ID
└ @ ERA5Reanalysis ~/work/ERA5Reanalysis.jl/ERA5Reanalysis.jl/src/variable/pressure.jl:263
julia> isPressure("vt",throw=false) # don't throw error, just show warning┌ Warning: 2023-06-03T04:20:26.577 - ERA5Reanalysis.jl - "vt" is not a valid PressureVariable identifier, use the function PressureCustom() to add this ERA5Variable to the list.
└ @ ERA5Reanalysis ~/work/ERA5Reanalysis.jl/ERA5Reanalysis.jl/src/variable/pressure.jl:233
false