Resetting the ERA5Variable Lists

We can reset the list of ERA5Variables back to the default predefined list using the function resetERA5Variables(). Alternatively, if we only want to reset SingleLevel or PressureLevel variables, we can use resetSingles() or resetPressures().

For the resetERA5Variables() and resetSingles() functions, the allfiles keyword will reset the list of SingleVariable variables as well, instead of just the list of SingleCustom variables.

ERA5Reanalysis.resetERA5Variables — Function

resetERA5Variables( allfiles :: Bool ) -> nothing

Reset the list of Single-Level and PressureCustom variables.

Arguments

allfiles : If false, only get rid of all the SingleCustom variables, but if true, then the SingleVariable list will be reset back to the default for ERA5Reanalysis.jl

ERA5Reanalysis.resetSingles — Function

resetSingles( allfiles :: Bool ) -> nothing

Reset the list of Single-Level variables.

Arguments

allfiles : If false, only get rid of all the SingleCustom variables, but if true, then the SingleVariable list will be reset back to the default for ERA5Reanalysis.jl

ERA5Reanalysis.resetPressures — Function

resetPressures() -> nothing

Reset the list of Pressure-Level variables to the ERA5Reanalysis default.