An Overview of the Download Functionality

The download functionality of ERA5Reanalysis.jl is built upon a Julia version of the CDSAPI (see here and here). In ERA5Reanalysis.jl, downloading a dataset can be as easy as download(ERA5Dataset,ERA5Variable,ERA5Region).

See below for the different methods. Note that downloading a PressureVariable is a bit more complicated and involved than downloading a SingleVariable because we need to also specify the range of pressures to download the data from the CDS when a PressureVariable is chosen.

Base.download — Method

download(
    e5ds :: Union{ERA5Hourly,ERA5Monthly},
    evar :: SingleVariable,
    ereg :: ERA5Region;
    ispy :: Bool = false,
    overwrite :: Bool = false
) -> nothing

Downloads ERA5 data from the CDS datastore for a specified Single-Level variable and geographic region. You can specify to download via python scripts generated by selecting ispy to true, or to instead use Julia directly.

You must have installed the CDSAPI on your machine and have accepted the terms and conditions on the CDS website in order for this to work.

Arguments

e5ds : The ERA5Dataset specified (Hourly or Monthly)

e5ds.start defines the start date

e5ds.stop defines the end date

e5ds.path defines the path to which all reanalysis data is saved

evar : Specifies the Single-Level variable to be downloaded

ereg : Specifies the GeoRegion and the resolution of the data to be downloaded

ipsy : Specifies whether to generate a python script that can be used to download the data instead of Julia

overwrite : false by default. If set to true, existing data will be overwritten.

Text

Base.download — Method

download(
    e5ds :: Union{ERA5Hourly,ERA5Monthly},
    evar :: Vector{SingleVariable},
    ereg :: ERA5Region;
    overwrite :: Bool = false
) -> nothing

Downloads ERA5 data from the CDS datastore for a specified variable and geographic region. There is no python option for this method.

You must have installed the CDSAPI on your machine and have accepted the terms and conditions on the CDS website in order for this to work.

Arguments

e5ds : The ERA5Dataset specified (Hourly or Monthly)

e5ds.start defines the start date

e5ds.stop defines the end date

e5ds.path defines the path to which all reanalysis data is saved

evar : Specifies a vector of Single-Level variables to be downloaded in a temporary file. Once downloaded, the different variables will be extracted out and saved in their respective folders.

ereg : Specifies the GeoRegion and the resolution of the data to be downloaded

overwrite : false by default. If set to true, existing data will be overwritten.

Text

Base.download — Method

download(
    e5ds :: Union{ERA5Hourly,ERA5Monthly},
    evar :: PressureVariable,
    ereg :: ERA5Region;
    ispy :: Bool = false,
    pall :: Bool = false,
    ptop :: Int = 0,
    pbot :: Int = 0,
    pvec :: Vector{Int} = [0],
    overwrite :: Bool = false
) -> nothing

Downloads ERA5 data from the CDS datastore for a specified Pressure-Level variable and geographic region. You can choose to specify (1) one pressure level, (2) a vector of pressure levels or (3) a set of pressure levels defined by their top and bottom levels. A python option is also available, but this will generate python scripts for all pressure levels.

You must have installed the CDSAPI on your machine and have accepted the terms and conditions on the CDS website in order for this to work.

Arguments

e5ds : The ERA5Dataset specified (Hourly or Monthly)

e5ds.start defines the start date

e5ds.stop defines the end date

e5ds.path defines the path to which all reanalysis data is saved

evar : Specifies a Pressure-Level variable to be downloaded. By default (if pall is not selected), then the pressure level closest to what is specified in evar.hPa will be downloaded.

ereg : Specifies the GeoRegion and the resolution of the data to be downloaded

ipsy : Specifies whether to generate a python script that can be used to download the data instead of Julia. If true, scripts for all pressure levels will be generated

pall : Indicates that we are selecting a range of pressure levels to be downloaded at the same time.

ptop : Indicates the pressure level at the top layer. ptop must be < pbot

pbot : Indicates the pressure level at the bottom layer. pbot must be > ptop

pvec : Defines a set of pressure levels to download. Overrides ptop and pbot

overwrite : false by default. If set to true, existing data will be overwritten.

TL;DR

The backend download functionality of ERA5Reanalysis.jl are based upon the build of (https://github.com/JuliaClimate/CDSAPI.jl)[CDSAPI.jl], but with the following extensions:

Eliminates the need to know the CDSAPI syntax for the frontend - all you need is to specify the Dataset, Variable and Region of interest

Extracts and places the downloaded data in a patterned, organized and systematic manner for easy retrieval

More detailed and organized logging information during the downloading process similar to the python version of CDSAPI

Allowing for repeated (up to 20) attempts at downloading a specific set of data

However, the download functionality of ERA5Reanalysis.jl is also limited in several ways:

It currently only is able to download the reanalysis data, not ensemble members

It currently is unable to retrieve any dataset outside the ERA5 Reanalysis datasets, including ERA5-Land data

It is not possible to specify multiple Pressure-Level variables for download in the same manner as Single-Level variables