The ERA5Module superType

ERA5 reanalysis data is stored on the Climate Data Store in several different categories, so different people with different needs may access different data types depending on their research. In ERA5Reanalysis, we defined these datasets as ERA5Module Types.

When defining an ERA5Module container, we also indicate the start and end dates of the dataset that we want to work on.

ERA5Reanalysis.ERA5Module — Type

ERA5Module

Abstract supertype for ERA5 dataset modules.

All ERA5Module Types contain the following fields:

modID : The module ID, that also acts as a prefix to filenames

eroot : The specified directory in which to save the data

dtbeg : The date for which downloads/analysis begins

dtend : The date for which downloads/analysis finishes

Note

The ERA5MonthlyHour subType also has the field hours that specifies the hour(s) of day for which monthly data is downloaded

Available Modules in ERA5Reanalysis.jl

The following datasets are currently supported by ERA5Reanalysis.jl:

Hourly reanalysis data (ERA5Hourly)

Monthly reanalysis data, which is further broken down in

Monthly-averaged data (ERA5Monthly)

Monthly-averaged hourly data (i.e. a monthly average of the diurnal cycle) (ERA5MonthlyHour)

ERA5Reanalysis.ERA5Hourly — Type

ERA5Hourly <: ERA5Module

Specifies that the dataset to be analyzed contains hourly data. All fields are the same as that specified in the ERA5Module docstring.

ERA5Reanalysis.ERA5Monthly — Type

ERA5Monthly <: ERA5Module

Specifies that the dataset to be analyzed contains monthly-mean data. All fields are the same as that specified in the ERA5Module docstring.

ERA5Reanalysis.ERA5MonthlyHour — Type

ERA5MonthlyHour <: ERA5Module

Specifies that the dataset to be analyzed contains monthly-mean hourly data. There is one additional field compared to ERA5Module:

Additional fields

hours : A vector of integers containing the hours-of-day to be downloaded/analyzed.

There are other potential modules that could be incorporated into ERA5Reanalysis.jl, such as

ERA5-Land data

Ensemble model averages, individual members, and standard deviations

Functions to create ERA5Modules

There are two functions used to create ERA5Modules:

ERA5Hourly(), which creates the ERA5Hourly subtype

ERA5Monthly(), which creates either the ERA5Monthly or ERA5MonthlyHour subtypes depending on inputs

ERA5Reanalysis.ERA5Hourly — Method

ERA5Hourly(;
    dtbeg :: TimeType,
    dtend :: TimeType,
    eroot :: AbstractString = homedir(),
) -> ERA5Hourly <: ERA5Module

A function that creates an ERA5Hourly module. All possible hours are downloaded, and data is saved month-by-month.

Keyword Arguments

eroot : The specified directory in which to save the data

dtbeg : The date for which downloads/analysis begins, automatically rounded to the nearest month

dtend : The date for which downloads/analysis finishes, automatically rounded to the nearest month

ERA5Reanalysis.ERA5Monthly — Method

ERA5Monthly(;
    dtbeg :: TimeType,
    dtend :: TimeType,
    eroot :: AbstractString = homedir(),
    hours :: Union{Int,Vector{Int}} = -1,
) -> ERA5Monthly <: ERA5Module or ERA5MonthlyHour <: ERA5Module

A function that creates an ERA5Monthly or ERA5MonthlyHour module depending on the input arguments of hours. Data is saved year-by-year.

Keyword Arguments

eroot : The specified directory in which to save the data

dtbeg : The date for which downloads/analysis begins, automatically rounded to the nearest year

dtend : The date for which downloads/analysis finishes, automatically rounded to the nearest year

hours : If hours < 0 or > 24, then it is assumed that we simply want monthly-averages without resolution on hour-of-day and we return an ERA5Monthly type. Otherwise, if an Int specifying a single hour-of-day (0-24), or a vector of Ints specifying multiple hours of days, then returns an ERA5MonthlyHour type.