Variable Types in ERA5Reanalysis.jl

In order to download data from the Climate Data Store, we need to specify at least a variable to download. This variable can be one found directly from CDS, or it can be a custom user-defined variable (in which case it has to be calculated by the user). Information regarding this variable will be loaded into an ERA5Variable.

ERA5Reanalysis.ERA5Variable — Type

ERA5Variable

Abstract supertype for ERA5 variables.

All ERA5Variable Types contain the following fields:

varID : The variable ID, that is also the identifier in the NetCDF files

lname : The variable long-name, which is used to specify retrievals from CDS

vname : The full-name of the variable

units : The units of the variable

SingleLevel and PressureLevel variables

There are two main ERA5Variable types in ERA5 reanalysis:

SingleLevel variables, such as surface temperature, or total cloud cover, that are provided in the (lon,lat) space

PressureLevel variables, such as atmospheric temperature, or specific humidity, that are provided in the (lon,lat,pressure) space

ERA5Reanalysis.SingleLevel — Type

SingleLevel <: ERA5Variable

Abstract supertype for Single-Level variables. Contains the following fields:

varID : The variable ID, that is also the identifier in the NetCDF files

lname : The variable long-name, which is used to specify retrievals from CDS

vname : The full-name of the variable

units : The units of the variable

dname : The name of the ERA5 dataset containing the variable

ERA5Reanalysis.PressureLevel — Type

PressureLevel <: ERA5Variable

Abstract supertype for Pressure-Level variables. Contains the following fields:

varID : The variable ID, that is also the identifier in the NetCDF files

lname : The variable long-name, which is used to specify retrievals from CDS

vname : The full-name of the variable

units : The units of the variable

hPa : The pressure-level height of the variable in concern

dname : The name of the ERA5 dataset containing the variable

Custom ERA5Variables

Each of these supertypes are further broken down into XXVariable and XXCustom subtypes, where XXVariable represents a variable that is available directly from CDS, and XXCustom variables are user-defined variables. XXCustom variables therefore cannot be downloaded - trying to do so will result in a method error - and can only be calculated from existing variable data.

ERA5Reanalysis.SingleVariable — Type

SingleVariable <: SingleLevel

Subtype for Single-Level variables that can be directly retrieved from the CDS

ERA5Reanalysis.SingleCustom — Type

SingleCustom <: SingleLevel

Subtype for custom user-defined Single-Level variables

ERA5Reanalysis.PressureVariable — Type

PressureVariable <: PressureLevel

Subtype for Pressure-Level variables that can be directly retrieved from the CDS

ERA5Reanalysis.PressureCustom — Type

PressureCustom <: PressureLevel

Subtype for custom user-defined Pressure-Level variables