Custom PressureLevel variables

In ERA5Reanalysis.jl, we have predefined all the pressure-level variables that are available directly from the CDS. As such, all custom PressureLevel variables are defined under the PressureCustom type.

Defining Custom PressureLevel Variables

PressureCustom variables are defined using the PressureVariable() function, as follows:

ERA5Reanalysis.PressureVariable — Type

PressureVariable(
    ST = String;
    varID :: AbstractString,
    lname :: AbstractString = "",
    vname :: AbstractString,
    units :: AbstractString,
    hPa   :: Int = 0,
    throw :: Bool = true
) -> evar :: PressureCustom

Create a custom Pressure-Level variable that is not in the default list exported by ERA5Reanalysis.jl. These variables are not available in the CDS store, and so they must be separately calculated from other variables and analyzed.

Keyword Arguments

varID : variable ID (in string format) used in the NetCDF file

lname : long-name for variable (used in specifying variable for CDS downloads)

vname : user-defined variable name

units : user-defined units of the variable

hPa : Pressure level specified in hPa. Default is 0, which indicates all levels.

throw : if hPa level does not exist and throw is true, throw error, otherwise find nearest pressure level

Removing Custom PressureLevel Variables

To remove a PressureCustom variable, we can use the rmPressure() function:

ERA5Reanalysis.rmPressure — Method

rmPressure( varID :: AbstractString ) -> nothing

Remove the Pressure-Level Variable with the ID varID from the lists.

Arguments

RegID : The keyword ID that will be used to identify the Pressure-Level Variable that is to be removed

An Example!

Let us define the custom variable "vt" for Virtual Temperature, which is a proxy to buoyancy in the atmosphere. Virtual Temperature is calculated from both the water vapour mixing ratio and atmospheric temperature and is given by the formula:

\[T_v = T(1+0.61q)\]

Virtual Temperature has the units "K", and for this example let us define the pressure level height we are interested in to be at 1000 hPa. The resultant PressureCustom variable is given by:

julia> using ERA5Reanalysis
julia> PressureVariable(
           varID = "vt",
           units = "K",
           hPa   = 1000,
           lname = "virtual_temperature",
           vname = "Virtual Temperature"
       )┌ Warning: 2022-11-01T04:31:13.741 - ERA5Reanalysis.jl - "vt" is not a valid PressureVariable identifier, use the function PressureCustom() to add this ERA5Variable to the list.
└ @ ERA5Reanalysis ~/work/ERA5Reanalysis.jl/ERA5Reanalysis.jl/src/variable/pressure.jl:236
[ Info: 2022-11-01T04:31:13.741 - ERA5Reanalysis.jl - Adding the PressureVariable "vt" to the list.
The Pressure-Level Variable "vt" has the following properties:
    Variable ID    (varID) : vt
    Long Name      (lname) : virtual_temperature
    Variable Name  (vname) : Virtual Temperature
    Variable Units (units) : K
    Pressure Level  (hPa)  : 1000
julia> tablePressures() ------------------ ------ ------------------------------------- ---------------------- -------------------------------------
   Variable Type      ID    Name                                         Units           ERA5 Long-Name
 ------------------ ------ ------------------------------------- ---------------------- -------------------------------------
  PressureVariable    u     Zonal Wind                                  m s**-1          u_component_of_wind
  PressureVariable    v     Meridional Wind                             m s**-1          v_component_of_wind
  PressureVariable    w     Vertical Wind                               Pa s**-1         lagrangian_tendency_of_air_pressure
  PressureVariable    t     Temperature                                    K             temperature
  PressureVariable    z     Pressure Level Geopotential                m**2 s**-2        geopotential
  PressureVariable    d     Divergence                                   s**-1           divergence
  PressureVariable    vo    Vorticity (Relative)                         s**-1           vorticity
  PressureVariable    pv    Potential Vorticity                   K m**-2 kg**-1 s**-1   potential_vorticity
  PressureVariable    r     Relative Humidity                              %             relative_humidity
  PressureVariable    q     Specific Humidity                          kg kg**-1         specific_humidity
  PressureVariable    cc    Cloud Cover Fraction                           %             fraction_of_cloud_cover
  PressureVariable   ciwc   Specific Cloud Ice Water Content           kg kg**-1         specific_cloud_ice_water_content
  PressureVariable   clwc   Specific Cloud Liquid Water Content        kg kg**-1         specific_cloud_liquid_water_content
  PressureVariable   crwc   Specific Rain Water Content                kg kg**-1         specific_rain_water_content
  PressureVariable   cswc   Specific Snow Water Contnet                kg kg**-1         specific_snow_water_content
  PressureVariable    o3    Ozone Mass Mixing Ration                   kg kg**-1         ozone_mass_mixing_ratio
   PressureCustom     vt    Virtual Temperature                            K             virtual_temperature
 ------------------ ------ ------------------------------------- ---------------------- -------------------------------------
julia> rmPressure("vt")ERROR: UndefKeywordError: keyword argument hPa not assigned
julia> isPressure("vt",throw=false) # don't throw error, just show warningtrue