Custom SingleLevel variables

The Climate Data Store provides a very large number of SingleLevel variables. As such, we only predefine a small subset of the variables available on CDS in ERA5Reanalysis.jl. However, only SingleVariable type variables can be downloaded from the Climate Data Store - SingleCustom type variables are not defined on CDS. Therefore, we allow the user the option to define if the custom SingleLevel variable they are defining exists on the Climate Data Store, and store it as a SingleVariable type.

Defining a new SingleVariable or SingleCustom

To define both SingleVariable and SingleCustom variables, we use the funciton SingleVariable(). To create a SingleCustom variable, the keyword argument inCDS must be set to false.

ERA5Reanalysis.SingleVariable — Type

SingleVariable(
    ST = String;
    varID :: AbstractString,
    lname :: AbstractString = "",
    vname :: AbstractString,
    units :: AbstractString,
    inCDS :: Bool = true
) -> evar :: SingleLevel

Create a custom Single-Level variable that is not in the default list exported by ERA5Reanalysis.jl. These variables are either available in the CDS store (whereby they can be both downloaded analyzed), or not (in which case means that they were separately calculated from other variables and analyzed).

Keyword Arguments

varID : variable ID (in string format) used in the NetCDF file

lname : long-name for variable (used in specifying variable for CDS downloads)

vname : user-defined variable name

units : user-defined units of the variable

inCDS : Boolean that indicates if this variable is available on the CDS store. True if available.

Removing Custom PressureLevel Variables

To remove a SingleLevel variable, we can use the rmSingle() function:

ERA5Reanalysis.rmSingle — Method

rmSingle( varID :: AbstractString ) -> nothing

Remove the Single-Level Variable with the ID varID from the lists.

Arguments

RegID : The keyword ID that will be used to identify the Single-Level Variable that is to be removed

An Example!

In this example, we add two SingleLevel variables to the list:

hvc, for High Vegetation Cover, which is a downloadable variable in the CDS

eke, for Eddy Kinetic Energy, which is not downloadable and must be calculated using other variables

The Eddy Kinetic Energy is calculated by the following:

\[E = \frac{1}{2g} \int_0^{p_s} \overline{u'^2} + \overline{v'^2} \>dp\]

And it is often used as a measure of storm-track intensity.

julia> using ERA5Reanalysis
julia> SingleVariable(
           varID = "cvh",
           units = "0-1",
           vname = "High Vegetation Cover",
           lname = "high_vegetation_cover",
       )The Single-Level Variable "cvh" has the following properties:
    Variable ID    (varID) : cvh
    Long Name      (lname) : high_vegetation_cover
    Variable Name  (vname) : High Vegetation Cover
    Variable Units (units) : 0-1
julia> SingleVariable(
           varID = "eke",
           units = "J m**-2",
           vname = "Eddy Kinetic Energy",
           lname = "eddy_kinetic_energy",
           inCDS = false
       )The Single-Level Variable "eke" has the following properties:
    Variable ID    (varID) : eke
    Long Name      (lname) : eddy_kinetic_energy
    Variable Name  (vname) : Eddy Kinetic Energy
    Variable Units (units) : J m**-2
julia> tableSingles() ---------------- ------ -------------------------------- ----------- ---------------------------------------
  Variable Type     ID    Name                               Units     ERA5 Long-Name
 ---------------- ------ -------------------------------- ----------- ---------------------------------------
  SingleVariable   u10    Zonal Wind at 10m Height          m s**-1    10m_u_component_of_wind
  SingleVariable   v10    Meridional Wind at 10m Height     m s**-1    10m_v_component_of_wind
  SingleVariable   u100   Zonal Wind at 100m Height         m s**-1    100m_u_component_of_wind
  SingleVariable   v100   Meridional Wind at 100m Height    m s**-1    100m_v_component_of_wind
  SingleVariable   t2m    2m Temperature                       K       2m_temperature
  SingleVariable   d2m    2m Dewpoint Temperature              K       2m_dewpoint_temperature
  SingleVariable   sst    Sea Surface Temperature              K       sea_surface_temperature
  SingleVariable   skt    Skin Temperature                     K       skin_temperature
  SingleVariable    sp    Surface Pressure                    hPa      surface_pressure
  SingleVariable   msl    Sea Level Pressure                  hPa      mean_sea_level_pressure
  SingleVariable    tp    Total Precipitation                  m       total_precipitation
  SingleVariable    cp    Convective Precipitation             m       convective_precipitation
  SingleVariable   lsp    Large-Scale Precipitation            m       large_scale_precipitation
  SingleVariable   tcw    Total Column Water               kg m**-2    total_column_water
  SingleVariable   tcwv   Total Column Water Vapour        kg m**-2    total_column_water_vapour
  SingleVariable   tcc    Total Cloud Cover                   0-1      total_cloud_cover
  SingleVariable   hcc    High Cloud Cover                    0-1      high_cloud_cover
  SingleVariable   mcc    Medium Cloud Cover                  0-1      medium_cloud_cover
  SingleVariable   lcc    Low Cloud Cover                     0-1      low_cloud_cover
  SingleVariable   cape   CAPE                             J kg**-1    convective_available_potential_energy
  SingleVariable   cin    CIN                              J kg**-1    convective_inhibition
  SingleVariable   tsr    Top Net Solar Radiation           J m**-2    top_net_solar_radiation
  SingleVariable   ttr    Top Net Thermal Radiation         J m**-2    top_net_thermal_radiation
  SingleVariable   ssr    Surface Net Solar Radiation        J m**-2   surface_net_solar_radiation
  SingleVariable   str    Surface Net Thermal Radiation     J m**-2    surface_net_thermal_radiation
  SingleVariable   sshf   Surface Sensible Heat Flux        J m**-2    surface_sensible_heat_flux
  SingleVariable   slhf   Surface Latent Heat Flux          J m**-2    surface_latent_heat_flux
  SingleVariable   cvh    High Vegetation Cover               0-1      high_vegetation_cover
   SingleCustom    eke    Eddy Kinetic Energy               J m**-2    eddy_kinetic_energy
 ---------------- ------ -------------------------------- ----------- ---------------------------------------
julia> rmSingle.(["eke","hvc"])[ Info: 2022-08-05T04:44:13.086 - ERA5Reanalysis.jl - Successfully removed the Single-Level variable defined by "eke"
┌ Warning: 2022-08-05T04:44:13.087 - ERA5Reanalysis.jl - "hvc" is not a valid SingleVariable identifier, use the function SingleVariable() or SingleCustom() to add this ERA5Variable to the list.
└ @ ERA5Reanalysis ~/work/ERA5Reanalysis.jl/ERA5Reanalysis.jl/src/variable/single.jl:200
┌ Warning: 2022-08-05T04:44:13.087 - ERA5Reanalysis.jl - No Single-Level variable defined by "hvc" exists, please make sure you specified the correct variable ID
└ @ ERA5Reanalysis ~/work/ERA5Reanalysis.jl/ERA5Reanalysis.jl/src/variable/single.jl:230
2-element Vector{Nothing}:
 nothing
 nothing
julia> isSingle("eke",throw=false) # don't throw error, just show warning┌ Warning: 2022-08-05T04:44:13.320 - ERA5Reanalysis.jl - "eke" is not a valid SingleVariable identifier, use the function SingleVariable() or SingleCustom() to add this ERA5Variable to the list.
└ @ ERA5Reanalysis ~/work/ERA5Reanalysis.jl/ERA5Reanalysis.jl/src/variable/single.jl:200
false