The Basic Components of ERA5Reanalysis.jl

There are three essential components to specifying an ERA5 reanalysis dataset:

The ERA5 module (hourly, monthly, month-by-hour, etc.), held in an ERA5Module supertype

The ERA5 variable (single-level, or pressure-level), held in the ERA5Variable supertype

The ERA5 region to be downloaded/analyzed, held in an ERA5Region supertype, built over a GeoRegion

The ERA5Module superType

ERA5 reanalysis data is stored on the Climate Data Store in several different categories, so different people with different needs may access different data types depending on their research. ERA5Reanalysis.jl distinguishes between the modules that are most often called using the ERA5Module superType:

Hourly reanalysis data

Monthly reanalysis data, which is further broken down in

Monthly-averaged data

Monthly-averaged hourly data (i.e. a monthly average of the diurnal cycle)

ERA5Reanalysis.ERA5Module — Type

ERA5Module

Abstract supertype for ERA5 dataset modules.

All ERA5Module Types contain the following fields:

modID : The module ID, that also acts as a prefix to filenames

eroot : The specified directory in which to save the data

dtbeg : The date for which downloads/analysis begins

dtend : The date for which downloads/analysis finishes

There are other potential modules that could be incorporated into ERA5Reanalysis.jl, such as

ERA5-Land data

Ensemble model averages, individual members, and standard deviations

More information on the different Types of ERA5Modules can be found here.

The ERA5Variable superType

There are two main variable types in ERA5 reanalysis:

Single-Level variables, such as surface temperature, or total cloud cover, that are provided in the (lon,lat) space

Pressure-Level variables, such as atmospheric temperature, or specific humidity, that are provided in the (lon,lat,pressure) space.

In ERA5Reanalysis.jl, information regarding a specific ERA5 variable will be loaded into an ERA5Variable Type, with more specifics given here

ERA5Reanalysis.ERA5Variable — Type

ERA5Variable

Abstract supertype for ERA5 variables.

All ERA5Variable Types contain the following fields:

varID : The variable ID, that is also the identifier in the NetCDF files

lname : The variable long-name, which is used to specify retrievals from CDS

vname : The full-name of the variable

units : The units of the variable

The ERA5Region superType

Finally, in order to specify an area for ERA5Reanalysis.jl to conduct downloads, or analyse data over, we need to specify an ERA5Region, which is built on top of the GeoRegion Type in GeoRegions.jl.

ERA5Reanalysis.ERA5Region — Type

ERA5Region

Structure that imports relevant GeoRegion properties used in the handling of the gridded ERA5 datasets.

All ERA5Region Types contain the following fields:

geo : The GeoRegion containing the geographical information

geoID : The ID used to specify the GeoRegion

gres : The resolution of the gridded data to be downloaded/analysed

fstr : String, for specification of folder and file name

isglb : A Bool, true if spans the globe, false if no