The Basic Components of ERA5Reanalysis.jl

There are three essential components to specifying an ERA5 reanalysis dataset:

The ERA5 module (hourly, monthly, month-by-hour, etc.), held in an ERA5Module supertype

The ERA5 variable (single-level, or pressure-level), held in the ERA5Variable supertype

The ERA5 region to be downloaded/analyzed, held in an ERA5Region supertype, built over a GeoRegion

The ERA5Variable superType

There are two main variable types in ERA5 reanalysis:

Single-Level variables, such as surface temperature, or total cloud cover, that are provided in the (lon,lat) space

Pressure-Level variables, such as atmospheric temperature, or specific humidity, that are provided in the (lon,lat,pressure) space.

In ERA5Reanalysis.jl, information regarding a specific ERA5 variable will be loaded into an ERA5Variable Type, with more specifics given here

ERA5Reanalysis.ERA5Variable — Type

ERA5Variable

Abstract supertype for ERA5 variables.

All ERA5Variable Types contain the following fields:

varID : The variable ID, that is also the identifier in the NetCDF files

lname : The variable long-name, which is used to specify retrievals from CDS

vname : The full-name of the variable

units : The units of the variable

The ERA5Region superType

Finally, in order to specify an area for ERA5Reanalysis.jl to conduct downloads, or analyse data over, we need to specify an ERA5Region, which is built on top of the GeoRegion Type in GeoRegions.jl.

ERA5Reanalysis.ERA5Region — Type

ERA5Region

Structure that imports relevant GeoRegion properties used in the handling of the gridded ERA5 datasets.

All ERA5Region Types contain the following fields:

geo : The GeoRegion containing the geographical information

geoID : The ID used to specify the GeoRegion

gres : The resolution of the gridded data to be downloaded/analysed

fstr : String, for specification of folder and file name

isglb : A Bool, true if spans the globe, false if no