Creating Custom ERA5Variables

While the list of available ERA5Variables in the Climate Data Store is very extensive, they are not exhaustive. There is therefore a need to define a list of custom ERA5Variables that will be otherwise calculated (although this has to be done separately outside of ERA5Reanalysis.jl), and then can thus be saved and manipulated and analyzed in the same way that the default ERA5Variables in ERA5Reanalysis.jl can be. In this page, we break this down into the SingleLevel and PressureLevel components

Custom SingleLevel Variables

As of ERA5Reanalysis.jl v0.3, we have predefined all the single-level variables that are available directly from the CDS. As such, all custom SingleLevel variables are defined under the SingleCustom type.

ERA5Reanalysis ≧ v0.3

All Single-Level variables defined in the Climate Data Store are now available in ERA5Reanalysis.jl, but only for versions ≧ v0.3.

Defining a new SingleCustom variable

To define aSingleCustom variables, we use the funciton SingleVariable().

ERA5Reanalysis.SingleVariable — Type

SingleVariable(
    ST = String;
    ID :: AbstractString,
    long :: AbstractString = "",
    name :: AbstractString,
    units :: AbstractString
) -> evar :: SingleLevel

Create a custom Single-Level variable that is not in the default list exported by ERA5Reanalysis.jl. These variables are not available in the CDS store, and so they must be separately calculated from other variables and analyzed.

Keyword Arguments

ID : variable ID (in string format) used in the NetCDF file

long : long-name for variable (used in specifying variable for CDS downloads)

name : user-defined variable name

units : user-defined units of the variable

inCDS : Boolean that indicates if this variable is available on the CDS store. True if available.

Removing Custom SingleLevel Variables

To remove a SingleCustom variable, we can use the rmSingle() function:

Missing docstring.

Missing docstring for rmSingle( varID :: AbstractString ). Check Documenter's build log for details.

An Example!

In this example, we add a SingleLevel variable to the list:

eke, for Eddy Kinetic Energy, which is not downloadable and must be calculated using other variables

The Eddy Kinetic Energy is calculated by the following:

\[E = \frac{1}{2g} \int_0^{p_s} \overline{u'^2} + \overline{v'^2} \>dp\]

And it is often used as a measure of storm-track intensity.

julia> using ERA5Reanalysis
julia> SingleVariable(
           ID = "eke",
           units = "J m**-2",
           name = "Eddy Kinetic Energy",
           long = "eddy_kinetic_energy"
       )┌ Warning: 2025-02-15T23:50:55.031 - ERA5Reanalysis.jl - "eke" is not a valid SingleVariable identifier, use the function SingleVariable() or SingleCustom() to add this ERA5Variable to the list.
└ @ ERA5Reanalysis ~/work/ERA5Reanalysis.jl/ERA5Reanalysis.jl/src/variable/is.jl:59
The Single-Level Variable "eke" has the following properties:
    Variable ID       (ID) : eke
    Long Name       (long) : eddy_kinetic_energy
    Variable Name   (name) : Eddy Kinetic Energy
    Variable Units (units) : J m**-2
julia> tableSingles(custom=true)ERROR: UndefVarError: `warn` not defined in `ERA5Reanalysis`
Suggestion: check for spelling errors or missing imports.
julia> rmSingle("eke")ERROR: UndefVarError: `rmSingle` not defined in `Main`
Suggestion: check for spelling errors or missing imports.
julia> isSingle("eke",throw=false) # don't throw error, just show warningERROR: UndefVarError: `isSingle` not defined in `Main`
Suggestion: check for spelling errors or missing imports.

Custom PressureLevel variables

In ERA5Reanalysis.jl, we have predefined all the pressure-level variables that are available directly from the CDS. As such, all custom PressureLevel variables are defined under the PressureCustom type.

Defining Custom PressureLevel Variables

PressureCustom variables are defined using the PressureVariable() function, as follows:

ERA5Reanalysis.PressureVariable — Type

PressureVariable(
    ST = String;
    ID :: AbstractString,
    long :: AbstractString = "",
    name :: AbstractString,
    units :: AbstractString,
    hPa   :: Int = 0,
    throw :: Bool = true
) -> evar :: PressureCustom

Create a custom Pressure-Level variable that is not in the default list exported by ERA5Reanalysis.jl. These variables are not available in the CDS store, and so they must be separately calculated from other variables and analyzed.

Keyword Arguments

ID : variable ID (in string format) used in the NetCDF file

long : long-name for variable (used in specifying variable for CDS downloads)

name : user-defined variable name

units : user-defined units of the variable

hPa : Pressure level specified in hPa. Default is 0, which indicates all levels.

throw : if hPa level does not exist and throw is true, throw error, otherwise find nearest pressure level

Removing Custom PressureLevel Variables

To remove a PressureCustom variable, we can use the rmPressure() function:

Missing docstring.

Missing docstring for rmPressure( varID :: AbstractString ). Check Documenter's build log for details.

An Example!

Let us define the custom variable "vt" for Virtual Temperature, which is a proxy to buoyancy in the atmosphere. Virtual Temperature is calculated from both the water vapour mixing ratio and atmospheric temperature and is given by the formula:

\[T_v = T(1+0.61q)\]

Virtual Temperature has the units "K", and for this example let us define the pressure level height we are interested in to be at 1000 hPa. The resultant PressureCustom variable is given by:

julia> using ERA5Reanalysis
julia> PressureVariable(
           ID = "vt",
           units = "K",
           hPa   = 1000,
           long = "virtual_temperature",
           name = "Virtual Temperature"
       )┌ Warning: 2025-02-15T23:50:56.732 - ERA5Reanalysis.jl - "vt" is not a valid PressureVariable identifier, use the function PressureCustom() to add this ERA5Variable to the list.
└ @ ERA5Reanalysis ~/work/ERA5Reanalysis.jl/ERA5Reanalysis.jl/src/variable/is.jl:110
The Pressure-Level Variable "vt" has the following properties:
    Variable ID       (ID) : vt
    Long Name       (long) : virtual_temperature
    Variable Name   (name) : Virtual Temperature
    Variable Units (units) : K
    Pressure Level   (hPa) : 1000
julia> tablePressures() ------ ------------------ ------------------------------------- ---------------------- -------------------------------------
   ID     Variable Type     Name                                         Units           ERA5 Long-Name
 ------ ------------------ ------------------------------------- ---------------------- -------------------------------------
   u     pressurevariable   Zonal Wind                                  m s**-1          u_component_of_wind
   v     pressurevariable   Meridional Wind                             m s**-1          v_component_of_wind
   w     pressurevariable   Vertical Wind                               Pa s**-1         vertical_velocity
   t     pressurevariable   Temperature                                    K             temperature
   z     pressurevariable   Pressure Level Geopotential                m**2 s**-2        geopotential
   d     pressurevariable   Divergence                                   s**-1           divergence
   vo    pressurevariable   Vorticity (Relative)                         s**-1           vorticity
   pv    pressurevariable   Potential Vorticity                   K m**-2 kg**-1 s**-1   potential_vorticity
   r     pressurevariable   Relative Humidity                              %             relative_humidity
   q     pressurevariable   Specific Humidity                          kg kg**-1         specific_humidity
   cc    pressurevariable   Cloud Cover Fraction                           %             fraction_of_cloud_cover
  ciwc   pressurevariable   Specific Cloud Ice Water Content           kg kg**-1         specific_cloud_ice_water_content
  clwc   pressurevariable   Specific Cloud Liquid Water Content        kg kg**-1         specific_cloud_liquid_water_content
  crwc   pressurevariable   Specific Rain Water Content                kg kg**-1         specific_rain_water_content
  cswc   pressurevariable   Specific Snow Water Contnet                kg kg**-1         specific_snow_water_content
   o3    pressurevariable   Ozone Mass Mixing Ration                   kg kg**-1         ozone_mass_mixing_ratio
 ------ ------------------ ------------------------------------- ---------------------- -------------------------------------
julia> rmPressure("vt")ERROR: UndefVarError: `rmPressure` not defined in `Main`
Suggestion: check for spelling errors or missing imports.
julia> isPressure("vt",throw=false) # don't throw error, just show warningERROR: UndefVarError: `isPressure` not defined in `Main`
Suggestion: check for spelling errors or missing imports.