Setting up the CDSAPI Key

In order for your downloads to work with ERA5Reanalysis, you first need to set up your CDSAPI key. To do this, you must first register with the Climate Data Store in order to obtain your key here.

Then, you can either follow the instructions above in the API-how-to, or you can simply use the function addCDSAPIkey() to do it for you if you don't want to fiddle around with hidden files and the like.

ERA5Reanalysis.addCDSAPIkey — Function

addCDSAPIkey(
    key :: AbstractString;
    url :: AbstractString = "https://cds.climate.copernicus.eu/api/v2",
    filename  :: AbstractString = ".cdsapirc",
    overwrite :: Bool = false
) -> nothing

Adds the user's CDSAPI key to a file in the homedir() (by default specified as .cdsapirc)

Arguments

key : The user's CDSAPI key

Keyword Arguments

url : The user's CDSAPI key

filename : The name of the file the url and key are saved to in the homedir()

overwrite : If true and if filename already exists, then overwrite

So, example

julia> using ERA5Reanalysis
julia> ckeys = ERA5Reanalysis.cdskey()[ Info: 2024-02-15T22:06:02.343 - CDSAPI - Loading CDSAPI credentials from /home/runner/.cdsapirc ...
Dict{Any, Any} with 2 entries:
  "key" => "199699:c52da207-6f7d-4ae8-bd33-085246faee6e"
  "url" => "https://cds.climate.copernicus.eu/api/v2"
julia> addCDSAPIkey("<your-key-here>")[ Info: 2024-02-15T22:06:02.546 - CDSAPI - Existing .cdsapirc file detected at /home/runner/.cdsapirc, since overwrite options is not selected, leaving file be ...  ...