The ERA5Dataset superType

ERA5 reanalysis data is stored on the Climate Data Store in several different categories, so different people with different needs may access different data types depending on their research. In ERA5Reanalysis, we defined these datasets as ERA5Dataset Types.

When defining an ERA5Dataset container, we also indicate the start and end dates of the dataset that we want to work on.

ERA5Reanalysis.ERA5Dataset — Type

ERA5Dataset

Abstract supertype for ERA5 reanalysis datasets.

All ERA5Dataset Types contain the following fields:

e5dID : The module ID, that also acts as a prefix to filenames

eroot : The specified directory in which to save the data

start : The date for which downloads/analysis begins

stop : The date for which downloads/analysis finishes

dtext : Is it the preliminary back extension from 1950-1978?

Note

The ERA5MonthlyHour subType also has the field hours that specifies the hour(s) of day for which monthly data is downloaded

Available Modules in ERA5Reanalysis.jl

The following datasets are currently supported by ERA5Reanalysis.jl:

Hourly reanalysis data (ERA5Hourly)

Monthly reanalysis data, which is further broken down in

Monthly-averaged data (ERA5Monthly)

Monthly-averaged hourly data (i.e. a monthly average of the diurnal cycle) (ERA5MonthlyHour)

ERA5Reanalysis.ERA5Hourly — Type

ERA5Hourly <: ERA5Dataset

Specifies that the dataset to be analyzed contains hourly data. All fields are the same as that specified in the ERA5Dataset docstring.

ERA5Reanalysis.ERA5Monthly — Type

ERA5Monthly <: ERA5Dataset

Specifies that the dataset to be analyzed contains monthly-mean data. All fields are the same as that specified in the ERA5Dataset docstring.

Missing docstring.

Missing docstring for ERA5MonthlyHour. Check Documenter's build log for details.

There are other potential modules that could be incorporated into ERA5Reanalysis.jl, such as

ERA5-Land data

Ensemble model averages, individual members, and standard deviations

Functions to create ERA5Datasets

There are two functions used to create ERA5Datasets:

ERA5Hourly(), which creates the ERA5Hourly subtype

ERA5Monthly(), which creates either the ERA5Monthly or ERA5MonthlyHour subtypes depending on inputs

ERA5Reanalysis.ERA5Hourly — Method

ERA5Hourly(;
    start :: TimeType,
    stop  :: TimeType,
    path  :: AbstractString = homedir(),
) -> ERA5Hourly <: ERA5Dataset

A function that creates an ERA5Hourly module. All possible hours are downloaded, and data is saved month-by-month.

Keyword Arguments

path : The specified directory in which to save the data

start : The date for which downloads/analysis begins, automatically rounded to the nearest month

stop : The date for which downloads/analysis finishes, automatically rounded to the nearest month

ERA5Reanalysis.ERA5Monthly — Method

ERA5Monthly(;
    start :: TimeType,
    stop  :: TimeType,
    path  :: AbstractString = homedir(),
    hours :: Bool = false,
) -> ERA5Monthly <: ERA5Dataset or ERA5MonthlyHour <: ERA5Dataset

A function that creates an ERA5Monthly or ERA5MonthlyHour module depending on the input arguments of hours. Data is saved year-by-year.

Keyword Arguments

path : The specified directory in which to save the data

start : The date for which downloads/analysis begins, automatically rounded to the nearest year

stop : The date for which downloads/analysis finishes, automatically rounded to the nearest year

hours : If false, download monthly-averaged data. If true, download monthly-averaged data for each hour